refactor(debug): extract TargetPathResolver from XMakeDebugSession
Lift the private `getTargetExecutable` (run `xmake l scripts/targetpath.lua
<target>` and parse the path between the __begin__/__end__ markers) into a
reusable `io.xmake.debug.TargetPathResolver`, so both the debug session and the
upcoming native run-config integration can resolve a target's built executable.
XMakeDebugSession now delegates to it; behavior is unchanged.
